+++ src/lib/kotlin/slatekit-jobs/src/main/kotlin/slatekit/jobs/Job.kt
@@ -1,6 +1,7 @@
 package slatekit.jobs
 
 import kotlinx.coroutines.*
+import kotlinx.coroutines.Job
 import kotlinx.coroutines.channels.Channel
 import slatekit.actors.*
 import slatekit.common.*
